chore: rename 'process.mts' to match its content
