+++ rt/src/Asserts.mts
@@ -1,7 +1,7 @@
 import { Thread, Capability } from './Thread.mjs';
 
 import { isListFlagSet, isTupleFlagSet } from './base/ValuesUtil.mjs';
-import * as proc from './base/process.mjs';
+import * as proc from './base/ProcessID.mjs';
 const ProcessID = proc.ProcessID;
 import { AbstractLevel } from './AbstractLevel.mjs';
 import { Level } from './Level.mjs';
