+++ rt/src/base/EqualityChecker.mts
@@ -1,7 +1,7 @@
 import {TroupeType} from './TroupeTypes.mjs'
 import {LVal} from './LVal.mjs'
 import { TroupeRawValue } from './TroupeRawValue.mjs';
-import * as proc from './process.mjs'
+import * as proc from './ProcessID.mjs'
 import * as levels  from '../Level.mjs'
 import { Level } from '../Level.mjs';
